Version 1.0.23 Turned peptide and protein IDs into longs as integers where in the internal database exceeded.

+++ src/main/java/org/rappsilber/fdr/entities/Peptide.java
@@ -49,7 +49,7 @@ public class Peptide extends AbstractFDRElement<Peptide>  { //implements Compara
     private HashMap<Protein,IntArrayList> m_positions = new HashMap<Protein, IntArrayList>();
     //int posSize = 0;
     /** unique ID for this peptide */
-    private int id;
+    private long id;
     /** assigned FDR for this peptide */
     public double m_fdr = -1;
     /** mass of this peptide */
@@ -66,7 +66,7 @@ public class Peptide extends AbstractFDRElement<Peptide>  { //implements Compara
      * @param pos position in the protein
      * @param length the length of the peptide
      */
-    public Peptide(int id, String sequence, boolean isDecoy, Protein protein, int pos, int length) {
+    public Peptide(long id, String sequence, boolean isDecoy, Protein protein, int pos, int length) {
         this(id, sequence, isDecoy, new Protein[]{protein}, new int[]{pos}, length);
     }
     
@@ -82,7 +82,7 @@ public class Peptide extends AbstractFDRElement<Peptide>  { //implements Compara
      * @param pos positions within the proteins
      * @param length the length of the peptide
      */
-    public Peptide(int id, String sequence, boolean isDecoy, Protein[] protein, int[] pos, int length) {
+    public Peptide(long id, String sequence, boolean isDecoy, Protein[] protein, int[] pos, int length) {
         this.sequence = sequence;
         this.isDecoy = isDecoy;
         this.length = length;
@@ -512,14 +512,14 @@ public class Peptide extends AbstractFDRElement<Peptide>  { //implements Compara
     /**
      * @return the unique peptide id
      */
-    public int getId() {
+    public long getId() {
         return id;
     }
 
     /**
      * @param id the peptide-id
      */
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
